LayoutTests/imported/w3c/web-platform-tests/streams/readable-streams/garbage-collection.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL garbage-collection.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
